+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   get any larger than specified (in mm). This is different from `--font-size`,
   which would force the text to be rendered at the same size, which risked
   causing overflow when rendering many varied text fields.
+- Fixed occasional issue where labels would rerender themselves to correct
+  minute scale differences (e.g. trying to scale down to correct a 99.9999999
+  undersize).
 
 # GFLabel 0.1.3 (2024-05-03)